backend/requirements.txt
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
psycopg==3.2.1
